+++ Sources/MacIMEKit/CJK.swift
@@ -8,44 +8,87 @@
 import Cocoa
 
 enum CJK {
-   /// Keep and reuse window/textField
    private static var window: NSWindow?
    private static var textField: NSTextField?
+   private static var delegate: NSObject?
 
-   /// Setup window
    private static func setup() {
-      guard window == nil, textField == nil else { return }
+      guard window == nil else { return }
 
       let w = NSWindow(
-         contentRect: NSRect(x: 0, y: 0, width: 10, height: 10),
-         styleMask: [.borderless],
+         contentRect: NSRect(x: 0, y: 0, width: 1, height: 1),
+         styleMask: [.borderless, .nonactivatingPanel],
          backing: .buffered,
          defer: false
       )
-      w.level = .screenSaver
-      w.backgroundColor = NSColor(white: 1.0, alpha: 0.5)
 
-      let tf = NSTextField(frame: NSRect(x: 0, y: 0, width: 10, height: 10))
-      tf.cell = NSSecureTextFieldCell() // password mode
+      w.backgroundColor = .clear
+      w.hasShadow = false
+      w.level = .screenSaver
+      w.isOpaque = false
+      w.alphaValue = 0.01
+
+      let effectView = NSVisualEffectView(frame: w.contentView!.bounds)
+      effectView.material = .popover
+      effectView.state = .active
+
+      let tf = NSTextField(frame: NSRect(x: 0, y: 0, width: 1, height: 1))
+
+      tf.font = NSFont.systemFont(ofSize: 13)
+      tf.bezelStyle = .roundedBezel
+      tf.isEditable = true
+      tf.isSelectable = true
+      tf.cell = NSSecureTextFieldCell()
+
+      class TextFieldDelegate: NSObject, NSTextFieldDelegate {
+         func controlTextDidChange(_ obj: Notification) {
+            _ = (obj.object as? NSTextField)?.stringValue
+         }
+
+         func control(
+            _: NSControl,
+            textView _: NSTextView,
+            doCommandBy commandSelector: Selector
+         ) -> Bool {
+            if commandSelector == #selector(NSResponder.cancelOperation(_:)) {
+               return true
+            }
+            return false
+         }
+      }
+
+      let d = TextFieldDelegate()
+      tf.delegate = d
+
+      w.contentView?.addSubview(effectView)
       w.contentView?.addSubview(tf)
+
+      w.initialFirstResponder = tf
+      w.setFrameOrigin(NSPoint(x: 0, y: 0))
 
       window = w
       textField = tf
+      delegate = d
    }
 
-   /// Refresh focus
    static func refresh() {
-      // setup if no window
       if window == nil || textField == nil {
          setup()
       }
 
       guard let w = window, let tf = textField else { return }
 
-      // Activate
-      NSApp.activate(ignoringOtherApps: true)
+      // NSApp.activate(ignoringOtherApps: true) // Disabled because it takes focus and does not return it in daemon.
       w.makeKeyAndOrderFront(nil)
       w.makeFirstResponder(tf)
+
+      // Run the loop once // TODO: NO NEED?
+      // RunLoop.main.run(until: Date().addingTimeInterval(0.02))
+
+      // Recieve events first
       tf.becomeFirstResponder()
+
+      // Hide window
+      w.orderOut(nil)
    }
 }